get rid of floats

+++ back/pochtron/models.py
@@ -3,7 +3,7 @@ from trade.models import Good
 
 
 class Alcohol(Good):
-    degree = models.FloatField()
-    volume = models.FloatField()
+    degree = models.DecimalField(max_digits=5, decimal_places=2)
+    volume = models.DecimalField(max_digits=6, decimal_places=3)
     def __str__(self):
         return self.name
